headtracked_parallax/requirements.txt
opencv-python
mediapipe
numpy
scipy
PySide6
PyOpenGL
pytest
